v0.2.1 hardening: CI-gated release, format v3, deserializer + race fixes, docs (#98–#104) (#105)
* v0.2.1 hardening: CI-gated release, format v3, deserializer + race fixes, docs (#98–#104)

- release.yml gates goreleaser on a full CI rerun (gate job: vet/lint/test -race)
  for the tagged SHA; a tag on a red commit produces no assets (#98).
- Fix flaky prefetch tests: fake S3 gains OnGetStart; tests synchronize on a GET
  in flight instead of sleeping. -race -count=20 clean (#98).
- goreleaser publishes stable-name raw binaries (lith_linux_{amd64,arm64})
  alongside the versioned archives; README/docs install uses the curl one-liner (#99).
- Docs currency: Zarr numbers reconciled (~35s vs ~25s, tier 1 shipped), command
  table + feature list to v0.2, "Why lith" note, knobs vs --help, SECURITY.md
  contact, mkdocs --strict (#100).
- Index format v3: offs/dirOffs uint32 -> uint64 (arena no longer ~4 GiB-capped;
  +4.0 B/key, ~4.6%; mmap zero-copy unchanged); v2 rejected with rebuild message.
  Deserializer bounds-checks every length/offset -> ErrCorruptIndex (never panics);
  fuzz test + committed regression corpus; 60s fuzz step in CI (#101).
- View.TotalSize summed at Root() and read as an immutable field; concurrent
  StatFs on a view is race-free (#102).
- CHANGELOG [0.2.1].

* docs: own the mountpoint before mounting (install-check finding #103)

The live 'Start here' check surfaced that 'sudo mkdir /mnt/x' leaves the
mountpoint root-owned, so fusermount3 refuses the user's mount ('user has no
write access to mountpoint'). Add 'sudo chown $USER' after the mkdir in the
Start-here tutorial and the README quick start.

+++ internal/index/build.go
@@ -129,7 +129,7 @@ func Build(entries []Entry, opts Options) *Index {
 	// 4. Pack the file arena and metadata; assign file inodes.
 	used := map[uint64]struct{}{0: {}, rootIno: {}}
 	n := len(final)
-	ix.offs = make([]uint32, n+1)
+	ix.offs = make([]uint64, n+1)
 	ix.sizes = make([]uint64, n)
 	ix.mtimes = make([]int64, n)
 	ix.etags = make([]uint64, n)
@@ -140,7 +140,7 @@ func Build(entries []Entry, opts Options) *Index {
 
 	var arena []byte
 	for i, e := range final {
-		ix.offs[i] = uint32(len(arena))
+		ix.offs[i] = uint64(len(arena))
 		arena = append(arena, e.Key...)
 		ix.sizes[i] = uint64(e.Size)
 		ix.mtimes[i] = e.MTime
@@ -148,7 +148,7 @@ func Build(entries []Entry, opts Options) *Index {
 		ix.inos[i] = ix.assignIno(hashKey(ix.prefix+e.Key), used)
 		accumulateDirs(dirMax, e.Key, e.MTime)
 	}
-	ix.offs[n] = uint32(len(arena))
+	ix.offs[n] = uint64(len(arena))
 	ix.arena = arena
 
 	// 5. Build the directory table: sorted paths, shared-namespace inodes, and
@@ -201,12 +201,12 @@ func (ix *Index) packDirs(dirMax map[string]int64, used map[uint64]struct{}) {
 	sort.Strings(dirs)
 
 	m := len(dirs)
-	ix.dirOffs = make([]uint32, m+1)
+	ix.dirOffs = make([]uint64, m+1)
 	ix.dirMtimes = make([]int64, m)
 	ix.dirInos = make([]uint64, m)
 	var arena []byte
 	for i, d := range dirs {
-		ix.dirOffs[i] = uint32(len(arena))
+		ix.dirOffs[i] = uint64(len(arena))
 		arena = append(arena, d...)
 		mt := dirMax[d]
 		if mt == 0 {
@@ -219,7 +219,7 @@ func (ix *Index) packDirs(dirMax map[string]int64, used map[uint64]struct{}) {
 		}
 		ix.dirInos[i] = ix.assignIno(hashKey(ix.prefix+d+"/"), used)
 	}
-	ix.dirOffs[m] = uint32(len(arena))
+	ix.dirOffs[m] = uint64(len(arena))
 	ix.dirArena = arena
 }
